Assistant checkpoint: 修复类型转换错误
Assistant generated file changes:
- src/main/java/com/hu/oneclick/server/service/impl/VersionEscapeAnalysisServiceImpl.java: 修复类型转换错误，统一使用double类型

---

User prompt:

为什么要转换来转去，用一种类型说法可以了。

+++ src/main/java/com/hu/oneclick/server/service/impl/VersionEscapeAnalysisServiceImpl.java
@@ -135,7 +135,7 @@ public class VersionEscapeAnalysisServiceImpl implements VersionEscapeAnalysisSe
 
         // 计算检测有效性
         double detectionEffectivenessValue = 100.0 - escapeRate.doubleValue();
-        stats.setDetectionEffectiveness(BigDecimal.valueOf(detectionEffectivenessValue));
+        stats.setDetectionEffectiveness(detectionEffectivenessValue);
 
         // 计算质量等级
         double escapeRateValue = escapeRate.doubleValue();
